elastic-update-service/requirements.txt
aiohttp 
requests
aiokafka
kafka-python